bazel: rename collapse_selects gazelle extension to collapse_deps_select
The extension only collapses the deps attribute, so the more specific name
better reflects its scope and matches the directive introduced in the next
commit.
